requirements.txt
# Flask requirements
Flask>=2.0
python-dotenv>=1.0.0

# Google Generative AI & Langchain Integration
google-generativeai>=0.4.0
langchain>=0.1.0
langchain-google-genai>=0.0.9
langchain-community>=0.0.20 # For SQLDatabase and other tools

# Vector Database
chromadb>=0.4.0

# Data Handling & Generation
pandas>=1.3
Faker>=15.0 # For data_generation

# Visualization (Optional - for data_generation scripts)
matplotlib>=3.5
seaborn>=0.11

# Add other dependencies as needed (e.g., gunicorn for deployment)
